A lot of changes!
In this you should find: 1) Sort of cleaned code, 2) Map with markers on the two stations with a red icon and black bicycle on it (which took for ever to make!!) 4) The date 5) The wordcountaddin 6) images 7) the url of our GitHub repository. 8) catchy title. 9) Footnotes 10) descriptions of the r code 11) date generator etc.